feat: 🎸 #6 增加screens  拆分頁面
拆分 登入頁面 首頁 與開放頁面
