+++ packages/splitbill-web/src/screens/HomeScreen/index.js
@@ -1,0 +1,3 @@
+export default function HomeScreen() {
+  return <h3>Home Screen</h3>;
+}
