+++ packages/splitbill-web/src/screens/PublicScreen/index.js
@@ -1,0 +1,3 @@
+export function PublicScreen() {
+  return <h3>Public Screen</h3>;
+}
